+++ src/app/modules/marketplace/services/modules-service/module.service.mock.ts
@@ -34,5 +34,3 @@ export const modulesSummaryList: Array<ModuleSummary> = [
         keywords: [],
     },
 ];
-
-export const accessToken = 'test-oidcToken';